utility functions are formed
