+++ psarev.alexandr/S1/unit.hpp
@@ -21,4 +21,5 @@ namespace psarev
     this->prev = prev;
   }
 }
+
 #endif
